+++ mycodo/config.py
@@ -14,7 +14,7 @@ sys.path.append(os.path.abspath(os.path.dirname(__file__)))
 from config_translations import TRANSLATIONS
 
 MYCODO_VERSION = '8.3.0'
-ALEMBIC_VERSION = 'a91e6a71028f'
+ALEMBIC_VERSION = 'dc9eddfc845d'
 
 #  FORCE_UPGRADE_MASTER
 #  Set True to enable upgrading to the master branch of the Mycodo repository.